A1Decider/requirements.txt
pysrt>=1.1.2
webvtt-py>=0.4.6
rich>=12.0.0
pynput>=1.7.6
tqdm>=4.64.0
spacy>=3.0.0

# Please download the SpaCy models separately:
# python -m spacy download de_core_news_lg
# python -m spacy download de_core_news_sm